Добавляю функцию для ивентлистенеров
