+++ css/style.css
@@ -926,7 +926,7 @@ fieldset {
   position: fixed;
   top: 0;
   left: 0;
-  z-index: 2;
+  z-index: 500;
 
   -webkit-box-sizing: border-box;
           box-sizing: border-box;
