+++ js/data.js
@@ -1,3 +1,5 @@
+import {getRandomIntInclusive, getRandomFloatInclusive, generateRandomValue, generateRandomList, shuffleArray} from './util.js';
+
 export const AVATAR_INDEX = {
   min: 1,
   max: 8,
@@ -67,7 +69,7 @@ export const SLICE = {
 
 //валидация формы
 
-export const TITLE_MIN_LENGTH = 30;
+export const TITLE_MIN_LENGTH = 3;
 
 export const TITLE_MAX_LENGTH = 100;
 
@@ -115,3 +117,37 @@ export const capacity = {
   2: [0, 1, 2],
   3: [3],
 };
+
+//Создание похожих карточек
+
+export const createListing = () => {
+  const location = {
+    lat: getRandomFloatInclusive(LOCATION_X.min, LOCATION_X.max, FLOAT_DIGITS),
+    lng: getRandomFloatInclusive(LOCATION_Y.min, LOCATION_Y.max, FLOAT_DIGITS),
+  };
+
+  return {
+    author: {
+      avatar: `img/avatars/user0${getRandomIntInclusive(AVATAR_INDEX.min, AVATAR_INDEX.max)}.png`,
+    },
+    offer: {
+      title: generateRandomValue(TITLES_LIST),
+      address: `${location.x}, ${location.y}`,
+      price: getRandomIntInclusive(PRICE.min, PRICE.max),
+      type: generateRandomValue(TYPES_LIST),
+      rooms: getRandomIntInclusive(ROOMS.min, ROOMS.max),
+      guests: getRandomIntInclusive(GUESTS.min, GUESTS.max),
+      checkin: generateRandomValue(CHECKIN_TIME_LIST),
+      checkout: generateRandomValue(CHECKIN_TIME_LIST),
+      features: generateRandomList(shuffleArray(FEATURES_LIST)),
+      description: generateRandomValue(DESCRIPTIONS_LIST),
+      photos: generateRandomList(shuffleArray(PHOTOS_LIST)),
+    },
+    location: {
+      lat: location.lat,
+      lng: location.lng,
+    },
+  }
+};
+
+export const similarListings = new Array(SIMILAR_LISTINGS_COUNT).fill(null).map(() => createListing());
