+++ js/form.js
@@ -1,11 +1,12 @@
 import {sendData} from './api.js';
-import {TITLE_MAX_LENGTH, TITLE_MIN_LENGTH, PRICE_MAX, pricesPerNight, capacity, Z_INDEX} from './data.js';
-import {onSuccess, onFormSend, onScreenClick, onEscClick, onErrorButtonClick} from './on-event.js';
+import {TITLE_MAX_LENGTH, TITLE_MIN_LENGTH, PRICE_MAX, pricesPerNight, capacity} from './data.js';
+import {onSuccess, onFormReset, showMessage} from './on-event.js';
 import {resetMainMarker} from './map.js';
 
 const mainBlock = document.querySelector('.promo');
 export const form = document.querySelector('.ad-form');
 const resetButton = form.querySelector('.ad-form__reset');
+export const submitButton = form.querySelector('ad-form__submit');
 const elementType = form.querySelector('#type');
 const elementPrice = form.querySelector('#price');
 const elementTimeIn = form.querySelector('#timein');
@@ -14,8 +15,10 @@ const elementTitle = form.querySelector('#title');
 const elementRooms = form.querySelector('#room_number');
 const elementGuests = form.querySelector('#capacity');
 export const addressField = form.querySelector('#address');
-const successMessage = document.querySelector('#success').content.querySelector('.success');
-const errorMessage = document.querySelector('#error').content.querySelector('.error');
+const successMessageTemplate = document.querySelector('#success').content.querySelector('.success');
+const successMessage = successMessageTemplate.cloneNode(true);
+const errorMessageTemplate = document.querySelector('#error').content.querySelector('.error');
+const errorMessage = errorMessageTemplate.cloneNode(true);
 export const errorButton = errorMessage.querySelector('.error__button');
 
 const validateRoomsGuests = function () {
@@ -85,21 +88,6 @@ const elementPriceInputHandler = function () {
   validatePrice();
 };
 
-const showSuccessMessage = function () {
-  mainBlock.appendChild(successMessage);
-  successMessage.style.zIndex = Z_INDEX;
-  onScreenClick(successMessage);
-  onEscClick(successMessage);
-};
-
-const showErrorMessage = function () {
-  mainBlock.appendChild(errorMessage);
-  errorMessage.style.zIndex = Z_INDEX;
-  onScreenClick(errorMessage);
-  onEscClick(errorMessage);
-  onErrorButtonClick(errorMessage);
-};
-
 elementType.addEventListener('change', elementTypeChangeHandler);
 elementTimeIn.addEventListener('change', elementTimeInChangeHandler);
 elementTimeOut.addEventListener('change', elementTimeOutChangeHandler);
@@ -112,7 +100,7 @@ resetButton.addEventListener('click', function (evt) {
   evt.preventDefault();
   form.reset();
   resetMainMarker();
-  onFormSend();
+  onFormReset();
 });
 
 form.addEventListener('submit', function (evt) {
@@ -121,9 +109,9 @@ form.addEventListener('submit', function (evt) {
 
   sendData(formData)
     .then(onSuccess)
+    .then(() => showMessage(mainBlock, successMessage))
+    .then(() => onFormReset())
     .then(() => form.reset())
     .then(() => resetMainMarker())
-    .then(() => onFormSend())
-    .then(() => showSuccessMessage())
-    .catch(() => showErrorMessage())
+    .catch(() => showMessage(mainBlock, errorMessage))
 });
